+++ src/core/mod.rs
@@ -1,4 +1,5 @@
 ﻿pub mod action_queue;
+pub mod context;
 pub mod dungeon;
 pub mod entity;
 pub mod error;